Support PCM -> fbank transform inside tensorflow
I'm attempting to export the model and use it in a non-python app.
While the tensorlow model is quite portable, the input pre-processing (audio.mfcc_fbank) is not.

This PR adds support to moving this stage into tensorflow by adding `pcm_input=True` when building the model.

When this is set, the input becomes a tensor with raw PCM data and shape [batch_size, num_samples].

The implementation attempts to be as close as possible to the original, however due to the different implementations/libraries, the results do differ a little bit.

The PR also provides an updated version of example_pcm.py, a modified version of example.py which uses raw PCM inputs and batched tensorflow operations